Version 1.0 functionality: Adding functionality to calculate and persist Student cumulative score in any CurriculumType weighted with progress.

+++ qpalx-elearning-domain/src/main/java/com/quaza/solutions/qpalx/elearning/domain/lms/adaptivelearning/statistics/StudentOverallProgressStatistics.java
@@ -110,7 +110,7 @@ public class StudentOverallProgressStatistics {
     }
 
     public double getTotalCurriculumCompletionPercent() {
-        int totalCurriculumItems = getTotalNumberOfCurriculumItems();
+        double totalCurriculumItems = getTotalNumberOfCurriculumItems();
         int totalAttemptedItems = getTotalNumberOfAttemptedItems();
 
         if (totalCurriculumItems > 0 && totalAttemptedItems > 0) {
@@ -121,8 +121,8 @@ public class StudentOverallProgressStatistics {
         return 0;
     }
 
-    public int getTotalNumberOfCurriculumItems() {
-        int totalCurriculumItems = totalNumberOfMicroLessons + totalNumberOfQuizzes + totalNumberOfQuestionBankItems;
+    public double getTotalNumberOfCurriculumItems() {
+        double totalCurriculumItems = totalNumberOfMicroLessons + totalNumberOfQuizzes + totalNumberOfQuestionBankItems;
         return totalCurriculumItems;
     }
